Land the round-16 B-batch findings
DoInfoViewProc spells the debug chord PADL2 | PADR2 (enum dropped),
the raw trig bits by name, the hidden menu id as an enumerator, and
its HUD coordinates in decimal; two self-echo comments trimmed.
GetNearestHumanoid's lone i = i + 1 folds to i++ (matched) and two
note claims are corrected (lhu narrowing; PAGE_CIVILIAN comparand,
not a threshold). FileOption names N_MUSIC_IDS across its three
coupled literals and joins the severed for-init (matched); the
caller-side TargetIO mask proved byte-required (retail's own — the
callee masks again). fast_tng4_ matches its leaf's radix and states
its one deviation from the recipe; launch_lightning_bolt_ drops a
no-op cast and describes the real flight direction; cd_init's header
sense fixed with the count spelled 10 - 1 (matched).

Whole image byte-identical.

+++ src/main.exe/launch_lightning_bolt_.c
@@ -48,7 +48,7 @@ void launch_lightning_bolt_(s16 frame)
         p.start.vx = start_pos->vx;
         p.start.vy = start_pos->vy;
         p.start.vz = start_pos->vz;
-        GetMoveSpeed(&move, dtR->vy, (s16)((rand() % 5) * 1000 + 4000), 0);
+        GetMoveSpeed(&move, dtR->vy, ((rand() % 5) * 1000 + 4000), 0);
         p.end.vx = start_pos->vx + move.vx;
         p.end.vy = Me_MOTION_C->target->locate.coord.t[1];
         p.end.vz = start_pos->vz + move.vz;
